[native] Move Mesh to Load
