+++ bve-native/src/lib.rs
@@ -27,9 +27,9 @@
 //! Structs are all prefixed with `BVE_` in C mode to help eliminate conflicts. C++ code is all in the `bve` namespace.
 //! Type names are kept as short as is reasonable, while still having clear connection with the underlying rust code.
 //!
-//! Free functions have their rust path loosely encoded in the name. For example [`bve::parse::mesh::mesh_from_str`] is
-//! [`bve_parse_mesh_from_string`](parse::mesh::bve_parse_mesh_from_string). Duplicate names are removed and idioms are
-//! changed to be comprehensible from the interface language.
+//! Free functions have their rust path loosely encoded in the name. For example
+//! [`bve::load::mesh::load_mesh_from_file`] is [`bve_load_mesh_from_file`](load::mesh::bve_load_mesh_from_file).
+//! Duplicate names are removed and idioms are changed to be comprehensible from the interface language.
 //!
 //! Free functions that are acting like member functions come in the form `BVE_Struct_Name_member_function_name`. They
 //! take their first argument by pointer.
@@ -86,7 +86,9 @@
 #![allow(clippy::wildcard_enum_match_arm)]
 
 pub mod filesystem;
+pub mod load;
 pub mod panic;
+pub mod parse;
 
 use std::borrow::Cow;
 use std::ffi::{CStr, CString};
@@ -231,5 +233,3 @@ unsafe fn unowned_ptr_to_str(input: &*const c_char) -> Cow<'_, str> {
 pub unsafe extern "C" fn bve_delete_string(ptr: *mut c_char) {
     CString::from_raw(ptr);
 }
-
-pub mod parse;
