+++ bve/src/load/mesh/mod.rs
@@ -4,6 +4,7 @@ use crate::parse::mesh::instructions::{create_instructions, post_process};
 use crate::parse::mesh::{BlendMode, FileType, Glow, GlowAttenuationMode, MeshError, MeshWarning};
 use crate::{ColorU8RGB, ColorU8RGBA};
 use cgmath::{Array, Vector2, Vector3};
+pub use execution::*;
 use indexmap::IndexSet;
 use std::ffi::OsStr;
 use std::path::Path;
